Report github.job
This has apparently always been available via the github context,
but I do not believe that the github context is naturally available via a file/environment variable.

So, here we expose it via an environment variable and then persist it for future steps.

https://github.com/github/docs/blame/35248ed2b23454f07f8c999e6e42fd664fedf04d/content/actions/learn-github-actions/contexts.md#L198

There are other places where the job id is calculated via an api call,
that could probably be simplified...